+++ src/Traits/Configuration/ComponentConfiguration.php
@@ -4,6 +4,7 @@ namespace Rappasoft\LaravelLivewireTables\Traits\Configuration;
 
 trait ComponentConfiguration
 {
+   
     public function setPrimaryKey(?string $key): self
     {
         $this->primaryKey = $key;
@@ -79,5 +80,16 @@ trait ComponentConfiguration
         $this->dataTableFingerprint = $dataTableFingerprint;
 
         return $this;
+    }
+
+    /**
+     * Returns a unique id for the table, used as an alias to identify one table from another session and query string to prevent conflicts
+     */
+    protected function generateDataTableFingerprint(): string
+    {
+        $className = str_split(static::class);
+        $crc32 = sprintf('%u', crc32(serialize($className)));
+
+        return base_convert($crc32, 10, 36);
     }
 }